+++ src/gql/interfaces/VizyBlockInterface.php
@@ -49,6 +49,7 @@ class VizyBlockInterface extends VizyNodeInterface
         return Craft::$app->getGql()->prepareFieldDefinitions(array_merge(parent::getFieldDefinitions(), [
             'enabled' => [
                 'name' => 'enabled',
+                'description' => 'Whether this Vizy block is enabled or not.',
                 'type' => Type::boolean(),
                 'resolve' => function($source, $arguments) {
                     return $source->getEnabled();
@@ -56,6 +57,7 @@ class VizyBlockInterface extends VizyNodeInterface
             ],
             'collapsed' => [
                 'name' => 'collapsed',
+                'description' => 'Whether this Vizy block is collapsed or not.',
                 'type' => Type::boolean(),
                 'resolve' => function($source, $arguments) {
                     return $source->attrs['collapsed'] ?? false;
@@ -63,6 +65,7 @@ class VizyBlockInterface extends VizyNodeInterface
             ],
             'blockTypeId' => [
                 'name' => 'blockTypeId',
+                'description' => 'The block type ID for this Vizy block.',
                 'type' => Type::string(),
                 'resolve' => function($source, $arguments) {
                     return $source->getBlockType()->id;
@@ -70,6 +73,7 @@ class VizyBlockInterface extends VizyNodeInterface
             ],
             'blockTypeHandle' => [
                 'name' => 'blockTypeHandle',
+                'description' => 'The block type handle for this Vizy block.',
                 'type' => Type::string(),
                 'resolve' => function($source, $arguments) {
                     return $source->getBlockType()->handle;