+++ src/gql/types/NodeCollectionType.php
@@ -3,9 +3,11 @@ namespace verbb\vizy\gql\types;
 
 use verbb\vizy\gql\interfaces\VizyNodeInterface;
 
+use Craft;
 use craft\gql\base\ObjectType;
 use craft\gql\GqlEntityRegistry;
 use craft\helpers\Gql;
+use craft\helpers\Json;
 
 use GraphQL\Type\Definition\ResolveInfo;
 use GraphQL\Type\Definition\Type;
@@ -30,13 +32,38 @@ class NodeCollectionType extends ObjectType
                 'fields' => [
                     'nodes' => [
                         'name' => 'nodes',
+                        'description' => 'Query nodes for this node collection.',
+                        'args' => [
+                            'where' => [
+                                'name' => 'where',
+                                'type' => Type::string(),
+                                'description' => 'Used to filter items based on params. This should be a JSON-encoded string.',
+                            ],
+                            'limit' => [
+                                'name' => 'limit',
+                                'type' => Type::int(),
+                                'description' => 'Limit the number of nodes returned.',
+                            ],
+                            'orderBy' => [
+                                'name' => 'orderBy',
+                                'type' => Type::string(),
+                                'description' => 'Return nodes ordered by a property.',
+                            ],
+                        ],
                         'type' => Type::listOf(VizyNodeInterface::getType($context)),
-                        'resolve' => function($source) {
-                            return $source->getNodes();
+                        'resolve' => function($source, $arguments) {
+                            if (isset($arguments['where'])) {
+                                $arguments['where'] = Json::decode($arguments['where']);
+                            }
+
+                            $query = Craft::configure($source->query(), $arguments);
+
+                            return $query->all();
                         },
                     ],
                     'rawNodes' => [
                         'name' => 'rawNodes',
+                        'description' => 'The raw JSON of nodes for this node collection.',
                         'type' => ArrayType::getType(),
                         'resolve' => function($source) {
                             return $source->getRawNodes();
@@ -44,6 +71,7 @@ class NodeCollectionType extends ObjectType
                     ],
                     'renderHtml' => [
                         'name' => 'renderHtml',
+                        'description' => 'The rendered HTML of nodes for this node collection.',
                         'type' => Type::string(),
                     ],
                 ],